chore(api): rename data_gouv by datagouv
