+++ packages/api/src/cli.ts
@@ -4,7 +4,7 @@ import 'dotenv/config' // see https://github.com/motdotla/dotenv#how-do-i-use-do
 import { mkdirSync, existsSync } from "fs";
 import LeCompteAssoCliController from "./modules/providers/leCompteAsso/interfaces/cli/leCompteAsso.cli.contoller";
 import MailNotifierCliController from "./modules/mail-notifier/interfaces/cli/mail-notifier.cli.controller";
-import DataGouvCliController from "./modules/providers/data_gouv/interfaces/cli/data_gouv.cli.controller";
+import DataGouvCliController from "./modules/providers/datagouv/interfaces/cli/datagouv.cli.controller";
 import FonjepCliController from "./modules/providers/fonjep/interfaces/cli/fonjep.cli.controller";
 import ChorusCliController from "./modules/providers/chorus/interfaces/cli/chorus.cli.controller";
 import OsirisCliController from "./modules/providers/osiris/interfaces/cli/osiris.cli.contoller";
