+++ packages/api/src/modules/providers/datagouv/interfaces/cli/datagouv.cli.controller.ts
@@ -3,8 +3,8 @@ import fs from "fs";
 import { StaticImplements } from "../../../../../decorators/staticImplements.decorator";
 import { CliStaticInterface} from "../../../../../@types/Cli.interface";
 import { findFiles } from "../../../../../shared/helpers/ParserHelper";
-import DataGouvParser from "../../data_gouv.parser";
-import dataGouvService from "../../data_gouv.service";
+import DataGouvParser from "../../datagouv.parser";
+import dataGouvService from "../../datagouv.service";
 import EntrepriseSirenEntity from "../../entities/EntrepriseSirenEntity";
 import RnaSiren from "../../../../rna-siren/entities/RnaSirenEntity";
 import { IStreamAction } from "../../@types";
